refactor: 생체 인증 로직 리팩터링 및 코루틴 기반으로의 전환
- `BiometricStateHolder` 및 `rememberBiometricAuthenticator`를 삭제하고 `FragmentActivity.authenticateBiometric` 확장 함수로 대체
- `BiometricPrompt` 콜백을 `suspendCancellableCoroutine`으로 래핑하여 비동기 흐름 제어 개선
- 인증 결과를 명확하게 처리하기 위해 `BiometricAuthResult` (Success, Canceled, Error) 인터페이스 도입
- 코루틴 취소 시 `cancelAuthentication`이 호출되도록 하여 리소스 누수 방지 로직 개선
- 사용자 취소(`ERROR_USER_CANCELED`, `ERROR_NEGATIVE_BUTTON`) 시 에러가 아닌 `Canceled` 상태를 반환하도록 수정
